pyproject.toml
[project]
name = "resume-customizer"
version = "0.1.0"
description = "AI-powered resume customization tool using Claude Code SDK"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chardet>=5.2.0",
    "claude-code-sdk>=0.0.13",
    "click>=8.2.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "toml>=0.10.2",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "--cov=resume_customizer",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--strict-markers",
    "-v"
]
asyncio_mode = "auto"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests"
]
